Move utils from pages
